create SettingsTable component
